+++ src/game/default/g_combat.h
@@ -29,9 +29,9 @@ _Bool G_CanDamage(const g_entity_t *targ, const g_entity_t *inflictor);
 vec3_t G_GetOrigin(const g_entity_t *ent);
 
 void G_Damage(g_entity_t *target, g_entity_t *inflictor, g_entity_t *attacker, const vec3_t dir,
-              const vec3_t point, const vec3_t normal, int16_t damage, int16_t knockback, uint32_t dflags,
+              const vec3_t point, const vec3_t normal, int32_t damage, int32_t knockback, int32_t dflags,
               g_mod_t mod);
 
-void G_RadiusDamage(g_entity_t *inflictor, g_entity_t *attacker, g_entity_t *ignore, int16_t damage,
-                    int16_t knockback, float radius, g_mod_t mod);
+void G_RadiusDamage(g_entity_t *inflictor, g_entity_t *attacker, g_entity_t *ignore, int32_t damage,
+                    int32_t knockback, float radius, g_mod_t mod);
 #endif /* __GAME_LOCAL_H__ */
